refactor: Notification컴포넌트를 화면 사이즈마다 다른 컴포넌트 렌더링하도록 변경
- 모바일에서는 Header가 필요한곳마다 렌더링되면서 unread-exsist api가 중복 호출되는 오류 발생
- Notification 컴포넌트를 화면 사이즈마다 선택적으로 렌더링하도록 하여 해결
