+++ src/components/common/Notification/Notification.tsx
@@ -12,9 +12,19 @@ import {
 import { ScrollArea } from '@/components/ui/scroll-area';
 import { useGetNewNotificationsCheck } from '@/hooks/notificationHooks/notificationHooks';
 
-const Notification = () => {
-  const isDesktop = useMediaQuery({ minWidth: 768 });
+const MobileNotification = () => {
   const router = useRouter();
+  const { data: hasNewNotificationData } = useGetNewNotificationsCheck();
+
+  return (
+    <BellIcon
+      isActive={hasNewNotificationData?.data?.hasUnread}
+      onClick={() => router.push('/notifications')}
+    />
+  );
+};
+
+const DesktopNotification = () => {
   const { data: hasNewNotificationData, refetch } =
     useGetNewNotificationsCheck();
   const pathname = usePathname();
@@ -25,25 +35,28 @@ const Notification = () => {
 
   return (
     <>
-      {isDesktop ? (
-        <Popover>
-          <PopoverTrigger>
-            <BellIcon isActive={hasNewNotificationData?.data?.hasUnread} />
-          </PopoverTrigger>
-          <PopoverContent>
-            <ScrollArea className='h-80'>
-              <NotificationList />
-            </ScrollArea>
-          </PopoverContent>
-        </Popover>
-      ) : (
-        <BellIcon
-          isActive={hasNewNotificationData?.data?.hasUnread}
-          onClick={() => router.push('/notifications')}
-        />
-      )}
+      <Popover>
+        <PopoverTrigger>
+          <BellIcon isActive={hasNewNotificationData?.data?.hasUnread} />
+        </PopoverTrigger>
+        <PopoverContent>
+          <ScrollArea className='h-80'>
+            <NotificationList />
+          </ScrollArea>
+        </PopoverContent>
+      </Popover>
     </>
   );
+};
+
+const Notification = () => {
+  const isDesktop = useMediaQuery({ minWidth: 768 });
+
+  if (isDesktop) {
+    return <DesktopNotification />;
+  }
+
+  return <MobileNotification />;
 };
 
 export default Notification;
